mine eiendeler, innstillingsknapp

+++ app/src/main/java/com/laanelitt/laanelittapp/objects/User.kt
@@ -1,7 +1,7 @@
 package com.laanelitt.laanelittapp.objects
 
 data class User(
-    val id: Double?,
+    val id: Int?,
     val firstName: String?,
     val lastName: String?,
     val profileImage: String?
